% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visGenome.R
\name{visGenome}
\alias{visGenome}
\title{Visualize Genome Properties}
\usage{
visGenome(data, seq_type, stats_type = "length_distr", ...)
}
\arguments{
\item{data}{a file path to a \code{fasta} file storing either DNA, CDS, or protein sequences.}

\item{seq_type}{sequence type stored in the \code{fasta} file specified in \code{data}.
Available options are:
\itemize{
\item \code{seq_type = "dna"}: the \code{fasta} file specified in \code{data} contains a \code{DNA} alphabet.
\item \code{seq_type = "cds"}: the \code{fasta} file specified in \code{data} contains coding sequences (CDS) (\code{DNA} alphabet).
\item \code{seq_type = "protein"}: the \code{fasta} file specified in \code{data} contains a \code{protein} alphabet.
}}

\item{stats_type}{a character string specifying the genome feature that shall be visualized: "length_distr", ... .}

\item{...}{additional parameters that shall be passed to the corresponding plot function.}
}
\value{
a plot object.
}
\description{
This function allows to visualize selected properties of input genomes such as:
\itemize{
\item gene length distribution: \code{stats_type} = \code{"length_distr"}

}
}
\examples{
\dontrun{
# download the proteome of Arabidopsis thaliana from refseq
# and store the corresponding proteome file in '_ncbi_downloads/proteomes'
# install.packages("biomartr")
Ath_proteome <- biomartr::getProteome( db = "refseq",
             organism = "Arabidopsis thaliana" )

# visualize the length distribution of the proteome of A. thaliana
visGenome(data = Ath_proteome,
          seq_type = "protein",
          stats_type = "length_distr", 
          main = "Histogram of A. thaliana protein sequence lengths",
          xlab = "Sequence Length",
          breaks = 100)
}
}
\author{
Hajk-Georg Drost
}
